requirements.txt
asgiref==3.4.1
Django==3.2.10
djangorestframework==3.12.4
pkg_resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
